src/ml-training/requirements.txt
# KDAP ML Training Requirements
# Azure ML SDK
azureml-core>=1.50.0
azureml-mlflow>=1.50.0
azure-ai-ml>=1.12.0
azure-identity>=1.15.0

# Core ML Libraries
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
polars>=0.19.0

# Feature Engineering
feature-engine>=1.6.0
category-encoders>=2.6.0

# Time Series
prophet>=1.1.0
statsmodels>=0.14.0

# Anomaly Detection
pyod>=1.1.0

# Model Interpretation
shap>=0.43.0
lime>=0.2.0

# Hyperparameter Tuning
optuna>=3.4.0
hyperopt>=0.2.7

# MLOps
mlflow>=2.9.0
evidently>=0.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0
